refactor(mobile): replace mobile BasicZIndexProvider with zIndexContext from core library
