+++ packages/core/client/src/schema-component/antd/action/index.tsx
@@ -8,6 +8,7 @@
  */
 
 export * from './Action';
+export * from './Action.Designer';
 export * from './ActionBar';
 export * from './context';
 export * from './hooks';
@@ -15,5 +16,5 @@ export * from './hooks/useGetAriaLabelOfAction';
 export * from './hooks/useGetAriaLabelOfDrawer';
 export * from './hooks/useGetAriaLabelOfModal';
 export * from './hooks/useGetAriaLabelOfPopover';
-export * from './Action.Designer';
 export * from './types';
+export * from './zIndexContext';
