+++ packages/plugins/@nocobase/plugin-mobile/src/client/adaptor-of-desktop/zIndex.ts
@@ -1,0 +1,11 @@
+/**
+ * This file is part of the NocoBase (R) project.
+ * Copyright (c) 2020-2024 NocoBase Co., Ltd.
+ * Authors: NocoBase Team.
+ *
+ * This project is dual-licensed under AGPL-3.0 and NocoBase Commercial License.
+ * For more information, please refer to: https://www.nocobase.com/agreement.
+ */
+
+// minimum z-index increment
+export const MIN_Z_INDEX_INCREMENT = 10;
