+++ packages/plugins/@nocobase/plugin-mobile/src/client/mobile/Mobile.tsx
@@ -18,12 +18,12 @@ import {
   OpenModeProvider,
   useAssociationFieldModeContext,
   usePlugin,
+  zIndexContext,
 } from '@nocobase/client';
 import React from 'react';
 import { isDesktop } from 'react-device-detect';
 
 import { ActionDrawerUsedInMobile, useToAdaptActionDrawerToMobile } from '../adaptor-of-desktop/ActionDrawer';
-import { BasicZIndexProvider } from '../adaptor-of-desktop/BasicZIndexProvider';
 import { useToAdaptFilterActionToMobile } from '../adaptor-of-desktop/FilterAction';
 import { InternalPopoverNesterUsedInMobile } from '../adaptor-of-desktop/InternalPopoverNester';
 import { MobileActionPage } from '../adaptor-of-desktop/mobile-action-page/MobileActionPage';
@@ -102,9 +102,9 @@ export const Mobile = () => {
                   <ResetSchemaOptionsProvider>
                     <AssociationFieldModeProvider modeToComponent={modeToComponent}>
                       {/* the z-index of all popups and subpages will be based on this value */}
-                      <BasicZIndexProvider basicZIndex={1000}>
+                      <zIndexContext.Provider value={1000}>
                         <MobileRouter />
-                      </BasicZIndexProvider>
+                      </zIndexContext.Provider>
                     </AssociationFieldModeProvider>
                   </ResetSchemaOptionsProvider>
                 </MobileAppProvider>
